refactor: rename CodeToolsPage and useCodeTools to CodeCliPage and useCodeCli
- Rename CodeToolsPage.tsx → CodeCliPage.tsx
- Rename useCodeTools.ts → useCodeCli.ts
- Rename useCodeTools.test.ts → useCodeCli.test.ts
- Rename getCodeToolsApiBaseUrl → getCodeCliApiBaseUrl
- Update all imports and references in routes, index, and tests
