+++ src/renderer/src/pages/code/index.ts
@@ -59,8 +59,8 @@ export const CLI_TOOL_PROVIDER_MAP: Record<string, (providers: Provider[]) => Pr
     providers.filter((p) => ['openai', 'openai-response', 'anthropic'].includes(p.type))
 }
 
-export const getCodeToolsApiBaseUrl = (model: Model, type: EndpointType) => {
-  const CODE_TOOLS_API_ENDPOINTS = {
+export const getCodeCliApiBaseUrl = (model: Model, type: EndpointType) => {
+  const CODE_CLI_API_ENDPOINTS = {
     aihubmix: {
       gemini: {
         api_base_url: 'https://aihubmix.com/gemini'
@@ -105,7 +105,7 @@ export const getCodeToolsApiBaseUrl = (model: Model, type: EndpointType) => {
 
   const provider = model.provider
 
-  return CODE_TOOLS_API_ENDPOINTS[provider]?.[type]?.api_base_url
+  return CODE_CLI_API_ENDPOINTS[provider]?.[type]?.api_base_url
 }
 
 // 解析环境变量字符串为对象
@@ -153,7 +153,7 @@ export const generateToolEnvironment = ({
   switch (tool) {
     case codeCLI.claudeCode:
       env.ANTHROPIC_BASE_URL =
-        getCodeToolsApiBaseUrl(model, 'anthropic') || modelProvider.anthropicApiHost || modelProvider.apiHost
+        getCodeCliApiBaseUrl(model, 'anthropic') || modelProvider.anthropicApiHost || modelProvider.apiHost
       env.ANTHROPIC_MODEL = model.id
       if (modelProvider.type === 'anthropic') {
         env.ANTHROPIC_API_KEY = apiKey
@@ -163,7 +163,7 @@ export const generateToolEnvironment = ({
       break
 
     case codeCLI.geminiCli: {
-      const apiBaseUrl = getCodeToolsApiBaseUrl(model, 'gemini') || modelProvider.apiHost
+      const apiBaseUrl = getCodeCliApiBaseUrl(model, 'gemini') || modelProvider.apiHost
       env.GEMINI_API_KEY = apiKey
       env.GEMINI_BASE_URL = apiBaseUrl
       env.GOOGLE_GEMINI_BASE_URL = apiBaseUrl
@@ -229,4 +229,4 @@ export const generateToolEnvironment = ({
   return { env }
 }
 
-export { default } from './CodeToolsPage'
+export { default } from './CodeCliPage'
